+++ bfg9000/build_inputs.py
@@ -19,7 +19,9 @@ def build_input(name):
 
 
 class Edge(object):
-    def __init__(self, build, output, final_output=None, extra_deps=None):
+    def __init__(self, build, output, final_output=None, extra_deps=None,
+                 description=None):
+        self.description = description
         self.raw_output = output
         self.output = listify(output)
         for i in self.output: