+++ CHANGES.md
@@ -12,6 +12,8 @@
   via bfg's logging system
 - `whole_archive()` now works with MSVC linkers
 - Add support for module-definition files when linking `.exe` or `.dll` files
+- Build steps now have a (customizable) friendly description when using the
+  Ninja backend
 
 ### Breaking changes
 - MSVC builds now automatically set `/EHsc` to improve standards-compliance and